pyproject.toml
[project]
name = "drf"
version = "0.1.0"
description = ""
authors = [
    {name = "Konstantin Voronin",email = "heavensmiting@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django (>=5.2.6,<6.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "ipython (>=9.5.0,<10.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.18.2"
black = "^25.9.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = '^(venv|tests)(/|$)'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''
[tool.isort]
# максимальная длина строки
line_length = 119

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=src --cov-report=term-missing"